# docs/contracts/openapi-v1-2026-02-19-phase5-dev.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "movement-whatsapp-automation-api",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/therapists": {
      "post": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "Create Therapist",
        "description": "Create a new therapist (with associated User record).\n\nCreates both User and Therapist records atomically.",
        "operationId": "create_therapist_api_v1_admin_therapists_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TherapistCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapistResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "List Therapists",
        "description": "List all therapists with basic info.",
        "operationId": "list_therapists_api_v1_admin_therapists_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TherapistListResponse"
                  },
                  "title": "Response List Therapists Api V1 Admin Therapists Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/therapists/{therapist_id}": {
      "get": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "Get Therapist",
        "description": "Get therapist detail with specialties.",
        "operationId": "get_therapist_api_v1_admin_therapists__therapist_id__get",
        "parameters": [
          {
            "name": "therapist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Therapist Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapistResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "Update Therapist",
        "description": "Update therapist details.",
        "operationId": "update_therapist_api_v1_admin_therapists__therapist_id__patch",
        "parameters": [
          {
            "name": "therapist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Therapist Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TherapistUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TherapistResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "Delete Therapist",
        "description": "Soft-delete therapist (set is_active=false).",
        "operationId": "delete_therapist_api_v1_admin_therapists__therapist_id__delete",
        "parameters": [
          {
            "name": "therapist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Therapist Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/therapists/{therapist_id}/specialties": {
      "post": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "Assign Specialty",
        "description": "Assign a specialty to a therapist.",
        "operationId": "assign_specialty_api_v1_admin_therapists__therapist_id__specialties_post",
        "parameters": [
          {
            "name": "therapist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Therapist Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpecialtyAssignment"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "List Therapist Specialties",
        "description": "List all specialties assigned to a therapist.",
        "operationId": "list_therapist_specialties_api_v1_admin_therapists__therapist_id__specialties_get",
        "parameters": [
          {
            "name": "therapist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Therapist Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SpecialtyResponse"
                  },
                  "title": "Response List Therapist Specialties Api V1 Admin Therapists  Therapist Id  Specialties Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/therapists/{therapist_id}/specialties/{specialty_id}": {
      "delete": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "Remove Specialty",
        "description": "Remove a specialty from a therapist.",
        "operationId": "remove_specialty_api_v1_admin_therapists__therapist_id__specialties__specialty_id__delete",
        "parameters": [
          {
            "name": "therapist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Therapist Id"
            }
          },
          {
            "name": "specialty_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Specialty Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/therapists/{therapist_id}/calendly-webhook/check": {
      "post": {
        "tags": [
          "Admin - Therapists"
        ],
        "summary": "Check Therapist Calendly Webhook",
        "description": "Admin-triggered Calendly webhook check for a therapist.\n\nUseful for operational diagnostics when onboarding spans multiple Calendly orgs.",
        "operationId": "check_therapist_calendly_webhook_api_v1_admin_therapists__therapist_id__calendly_webhook_check_post",
        "parameters": [
          {
            "name": "therapist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Therapist Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendlyWebhookCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalendlyWebhookCheckResponse": {
        "properties": {
          "endpoint_url": {
            "type": "string",
            "title": "Endpoint Url"
          },
          "user_uri": {
            "type": "string",
            "title": "User Uri"
          },
          "organization_uri": {
            "type": "string",
            "title": "Organization Uri"
          },
          "has_matching_webhook": {
            "type": "boolean",
            "title": "Has Matching Webhook"
          },
          "needs_registration": {
            "type": "boolean",
            "title": "Needs Registration"
          },
          "missing_events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Events",
            "default": []
          },
          "subscriptions": {
            "items": {
              "$ref": "#/components/schemas/CalendlyWebhookSubscriptionInfo"
            },
            "type": "array",
            "title": "Subscriptions",
            "default": []
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "endpoint_url",
          "user_uri",
          "organization_uri",
          "has_matching_webhook",
          "needs_registration"
        ],
        "title": "CalendlyWebhookCheckResponse",
        "description": "Response for webhook registration check."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SpecialtyAssignment": {
        "properties": {
          "specialty_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Specialty Id"
          }
        },
        "type": "object",
        "required": [
          "specialty_id"
        ],
        "title": "SpecialtyAssignment",
        "description": "Request schema for assigning specialty to therapist."
      },
...
      "SpecialtyResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description",
          "is_active"
        ],
        "title": "SpecialtyResponse",
        "description": "Response schema for specialty."
      },
...
      "TherapistCreate": {
        "properties": {
          "neon_auth_sub": {
            "type": "string",
            "minLength": 1,
            "title": "Neon Auth Sub",
            "description": "Neon Auth subject ID"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email",
            "description": "Therapist email (unique)"
          },
          "display_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Display Name"
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number"
          },
          "calendly_user_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendly User Uri",
            "description": "Calendly user URI"
          }
        },
        "type": "object",
        "required": [
          "neon_auth_sub",
          "email",
          "display_name"
        ],
        "title": "TherapistCreate",
        "description": "Request schema for creating therapist.",
        "example": {
          "calendly_user_uri": "https://api.calendly.com/users/XXXXX",
          "display_name": "Dr. Smith",
          "email": "dr.smith@clinic.com",
          "license_number": "PT203315",
          "neon_auth_sub": "auth-therapist-123"
        }
      },
...
      "TherapistListResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "specialty_count": {
            "type": "integer",
            "title": "Specialty Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name",
          "license_number",
          "is_active",
          "email",
          "specialty_count"
        ],
        "title": "TherapistListResponse",
        "description": "Response schema for list of therapists (lightweight)."
      },
...
      "TherapistResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "calendly_user_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendly User Uri"
          },
          "specialties": {
            "items": {
              "$ref": "#/components/schemas/SpecialtyResponse"
            },
            "type": "array",
            "title": "Specialties"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "display_name",
          "license_number",
          "is_active",
          "calendly_user_uri",
          "specialties",
          "created_at"
        ],
        "title": "TherapistResponse",
        "description": "Response schema for therapist with full details."
      },
      "TherapistUpdate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "license_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "License Number"
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "calendly_user_uri": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calendly User Uri"
          }
        },
        "type": "object",
        "title": "TherapistUpdate",
        "description": "Request schema for updating therapist."
      },
...
    }
  }
}